+++ client/src/components/CheckoutForm.js
@@ -1,4 +1,7 @@
-import React, { useState } from "react";
+import React from "react";
+
+//Import useForm custom hook
+import { useForm } from '../hooks/useForm';
 
 const initialValue = {
   firstName: "",
@@ -14,17 +17,23 @@ const initialValue = {
 // and replace the necessary stateful logic from CheckoutForm with the hook
 
 const CheckoutForm = (props) => {
-  const [showSuccessMessage, setShowSuccessMessage] = useState(false);
-  const [values, setValues] = useState(initialValue);
+  
+  //Moved these over to useForm hook inside of custom hook
+  // const [showSuccessMessage, setShowSuccessMessage] = useState(false);
+  // const [values, setValues] = useState(initialValue);
 
-  const handleChanges = (e) => {
-    setValues({ ...values, [e.target.name]: e.target.value });
-  };
+  // const handleChanges = (e) => {
+  //   setValues({ ...values, [e.target.name]: e.target.value });
+  // };
 
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setShowSuccessMessage(true);
-  };
+  // const handleSubmit = (e) => {
+  //   e.preventDefault();
+  //   setShowSuccessMessage(true);
+  // };
+
+  const [showSuccessMessage, values, handleChanges, handleSubmit] = useForm(initialValue);
+
+  
 
   return (
     <>
